desktop/resources/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
sqlalchemy>=2.0.23
alembic>=1.13.1
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.9
pillow>=10.2.0
pytesseract>=0.3.10
weasyprint>=61.0
